add: 授業

+++ 2025_st42_r04_surfaceview/app/src/main/java/com/example/a2025_st42_r04_surfaceview/BallObj.java
@@ -17,7 +17,7 @@ public class BallObj {
             dy = -dy;
         }
 
-        ballX += dx;
-        ballY += dy;
+        ballX += (int) dx;
+        ballY += (int) dy;
     }
 }
